requirements.txt
pandas
matplotlib
numpy
Pillow
reportlab
ipython
opencv-python
h5py
pdf2image